switcher: also set the language cookie when the current value is wrong.
I am not completely sure this can happen in practice, but I have seen some unwanted behavior.
That seems to have more to do with the redirect being a 301 (permanent),
which is fine normally, but is not always nice in development.
Clearing the (Firefox) browser cache also clears the stored permanent redirect though,
which can help when you find yourself redirected to the English folder while still having the Dutch language in the cookie.
It does not look like the change actually helps for this, but it is still logical.