Make search look in "Lanuguage Shared" too.
When using the search viewlet, or the search form, the path
parameter is a string, not a dictionary having a "query" key with
a list of paths as a value.

We now check first to see if query['path'] is a string to make the
search form work on the /shared folder also.