Robot tests: first wait until page contains element.
Only after this, wait until element is visible.
Otherwise you get an error:
Message: Element not found in the cache - perhaps the page has changed since it was looked up
If the element was already on the page, this is of course not needed.
But here the element was added dynamically, so we need this extra step.
At least for Firefox.