Fix "Insufficient Privileges" error when adding translations.
In some cases, e.g. when the target folder is "private", the traversal
does not work well with the URL redirection. Luckily, looking up
the `add_at_translation` view and using it directly fixes this.
Kudos to Mathias Leimgruber (@maethu) for suggesting this[1].

The underlying problem seems to be the two `hasattr` calls in
`BaseRequest.traverse()` [2], although `getRoles` a few lines below
apparently works correctly.  In any case, if the current user has
no access to the target folder, return the redirection URL as a
string causes the following error:

  Traceback (innermost last):
    Module ZPublisher.Publish, line 127, in publish
    Module ZPublisher.BaseRequest, line 623, in traverse
    Module ZPublisher.HTTPResponse, line 756, in unauthorized
  Unauthorized: You are not authorized to access this resource.
  No Authorization header found.

[1] https://stackoverflow.com/q/23109624.
[2] https://github.com/zopefoundation/Zope/blob/2.13.28/src/ZPublisher/BaseRequest.py#L509